altered the model attributes

+++ restaurant/models.py
@@ -12,8 +12,8 @@ class Booking(models.Model):
     
 class Menu(models.Model):
     title = models.CharField(max_length=255, verbose_name='Item Name')
-    price = models.DecimalField(max_digits=10, decimal_places=2, verbose_name='Price $')
-    inventory = models.IntegerField()
+    price = models.DecimalField(max_digits=6, decimal_places=2, verbose_name='Price $')
+    inventory = models.SmallIntegerField()
     
     def __str__(self):
         return f'{self.title}: ${self.price}'
